+++ example/metro.config.js
@@ -1,7 +1,6 @@
 const path = require('path');
 const { getDefaultConfig, mergeConfig } = require('@react-native/metro-config');
 const escape = require('escape-string-regexp');
-const exclusionList = require('metro-config/src/defaults/exclusionList');
 const pak = require('../package.json');
 
 const root = path.resolve(__dirname, '..');
@@ -20,10 +19,8 @@ const peers = Object.keys({ ...pak.peerDependencies });
 const config = {
   watchFolders: [root],
   resolver: {
-    blockList: exclusionList(
-      peers.map(
-        (m) => new RegExp(`^${escape(path.join(root, 'node_modules', m))}\\/.*$`)
-      )
+    blockList: peers.map(
+      (m) => new RegExp(`^${escape(path.join(root, 'node_modules', m))}\\/.*$`)
     ),
     extraNodeModules: peers.reduce((acc, name) => {
       acc[name] = path.join(__dirname, 'node_modules', name);